post comment/uncomment real time update done

+++ app/Http/Controllers/CommentController.php
@@ -1,8 +1,10 @@
 <?php
-
 namespace App\Http\Controllers;
 
+use App\Events\PostCommentUpdated;
+use App\Http\Controllers\Controller;
 use App\Models\Comment;
+use App\Models\Post;
 use Illuminate\Http\Request;
 
 class CommentController extends Controller
@@ -18,13 +20,18 @@ class CommentController extends Controller
             'comment' => 'required',
         ]);
 
-        $comment = new Comment;
-
+        $comment          = new Comment;
         $comment->post_id = $request->input('post_id');
         $comment->user_id = $request->input('user_id');
-        $comment->text = $request->input('comment');
+        $comment->text    = $request->input('comment');
 
         $comment->save();
+
+        // --- REAL TIME TRIGGER ---
+        $post = Post::find($request->input('post_id'));
+        if ($post) {
+            broadcast(new PostCommentUpdated($post))->toOthers();
+        }
     }
 
     /**
@@ -33,6 +40,16 @@ class CommentController extends Controller
     public function destroy($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
+
+        if ($comment) {
+            $post = Post::find($comment->post_id);
+            $comment->delete();
+
+            // --- REAL TIME TRIGGER ---
+
+            if ($post) {
+                broadcast(new PostCommentUpdated($post))->toOthers();
+            }
+        }
     }
 }
